HOLY HECK I GOT DETAILS SHELL TO SHOW
Summary: FINALLY GOT DETAILS SHELL TO SHOW WHEN CLICKING ON ITEM VIEW (sorry I'm v excited omg)

Test Plan: {F133988715}

Reviewers: j053, madjangba, kychow

Reviewed By: kychow

Differential Revision: https://phabricator.intern.facebook.com/D9022049

Tasks: T32017548, T31834094

Signature: 9022049:1532723463:b0800bdd86a5fb59b6898fb936d90c59b104bf4d

+++ app/src/main/java/com/kychow/jayjoska/ItineraryAdapter.java
@@ -40,7 +40,7 @@ public class ItineraryAdapter extends RecyclerView.Adapter<ItineraryAdapter.View
     public void onBindViewHolder(@NonNull ViewHolder holder, int position) {
         Place place = mItinerary.get(position);
         holder.mName.setText(place.getName());
-        holder.mDistance.setText(Long.toString(place.getDistance()));
+        holder.mDistance.setText(Double.toString(place.getDistance()));
         holder.mCategory.setText(place.getCategory());
         holder.mRating.setRating(place.getRating());
     }
